api/code/requirements.txt
flask
gunicorn
prometheus-flask-exporter
PyYAML
kopf
kubernetes
redis